// src/generated/client/models/VacationType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Timebank
 * Timebank API documentation
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Enum for vacation type
 * @export
 */
export const VacationType = {
    VACATION: 'VACATION',
    UNPAID_TIME_OFF: 'UNPAID_TIME_OFF',
    SICKNESS: 'SICKNESS',
    PERSONAL_DAYS: 'PERSONAL_DAYS',
    MATERNITY_PATERNITY: 'MATERNITY_PATERNITY',
    CHILD_SICKNESS: 'CHILD_SICKNESS'
} as const;
export type VacationType = typeof VacationType[keyof typeof VacationType];


export function VacationTypeFromJSON(json: any): VacationType {
    return VacationTypeFromJSONTyped(json, false);
}

export function VacationTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): VacationType {
    return json as VacationType;
}

export function VacationTypeToJSON(value?: VacationType | null): any {
    return value as any;
}